requirements.txt
# Web scraping
requests==2.28.2
beautifulsoup4==4.11.2
selenium==4.8.2
webdriver-manager==3.8.5
lxml==4.9.2

# Data processing
pandas==1.5.3
numpy==1.24.2
scipy==1.10.1

# Natural language processing
nltk==3.8.1
spacy==3.5.1
textblob==0.17.1

# Image processing
Pillow==9.4.0

# Network utilities
dnspython==2.3.0

# Progress reporting
tqdm==4.65.0

# Visualization (for any backend analysis)
matplotlib==3.7.1
seaborn==0.12.2

# Utility
python-dotenv==1.0.0 